requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==3.0.4
click==6.7
Django==3.2.5
django-cors-headers==3.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==2.7
itsdangerous==0.24
Jinja2==2.11.3
MarkupSafe==0.23
nulltype==2.3.1
plaid-python==3.4.0
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.20.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.24.3
Werkzeug==2.0.1
gunicorn==20.0.4
